refactor(studio): split ScoreGauge into consts and utils modules [ASTD-310]
Move the score-tier constants and gauge geometry into consts.ts and the
pure helper functions into utils.ts. index.tsx keeps only the component and
re-exports both modules so existing import paths are unchanged.
